requirements.txt
python-telegram-bot==20.7
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1